Finished project
Have finished all of the steps.
The after I made sure the npm test came back as passed.
Made personal changes to the game.
Changed the following:
Background of Asgard
Moles were turned into Loki
Figured out how to enable sound and added a fun sountrack when the game starts.
The audio restarts when the game is started again.
Now there is a hit sound when the "mole" is whacked.
Changed the H1 to Whack-a-Loki.
Changed the H1 color to a bluish color to represent the lightning from Thor.
Added a select/drop down menu to choose difficulty BEFORE and AFTER a game. It defaults to HARD when no selection is picked.